openmemory/api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.5
fastapi-pagination>=0.12.0
mem0ai>=0.1.92
openai>=1.40.0
mcp[cli]>=1.3.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0
tenacity==9.1.2
anthropic==0.51.0
ollama==0.4.8
chromadb>=0.5.0
elasticsearch>=8.0.0,<9.0.0
faiss-cpu>=1.10.0
opensearch-py>=2.4.0
pymilvus>=2.4.0
redis>=5.0.0
redisvl>=0.3.0
weaviate-client>=4.0.0
